fix(lint): clean mechanical and fork-only worker findings
Unwrap always-true lastActiveTask nil guard (SA4031), flatten
DecrConnErrors else-if (gocritic), and apply non-semantic
Deprecated/errors.Is/As/cast/t.Fatalf cleanups across fork-only
test and production code.

- progress.go: DEPRECATED → Deprecated godoc directive (4 sites)
- worker.go: remove always-true nil guard on lastActiveTask;
  flatten nested else-if for DecrConnErrors steal-skip
- handlepause_vp_guard_test.go: errors.Is for ErrPaused
- pause_regression_test.go: errors.Is for ErrPaused (2 sites)
- preallocate_windows_test.go: errors.As for windows.Errno,
  errors.Is for io.EOF
- rate_limiter_test.go: errors.Is for context errors
- scale_prewarm_test.go: t.Fatal(fmt.Sprintf) → t.Fatalf
- buffer_pool.go: remove redundant bufferTier() cast
- vp_overcount_regression_test.go: remove redundant int64() cast

+++ internal/surge/strategy/concurrent/vp_overcount_regression_test.go
@@ -476,7 +476,7 @@ func TestVPOvercount_DownloadTaskCountClampedMultiChunk(t *testing.T) {
 
 	// Two chunks: chunk 0 = [0, 128KiB), chunk 1 = [128KiB, 256KiB).
 	chunkSize := int64(128 * utils.KiB)
-	fileSize := int64(2 * chunkSize)
+	fileSize := 2 * chunkSize
 	handler := &fullRangeHandler{fileSize: fileSize}
 	srv := testutil.NewHTTPServerT(t, handler)
 	t.Cleanup(srv.Close)
